Conflicts:
	crawl-ref/settings/messages.txt
	crawl-ref/source/abl-show.cc
	crawl-ref/source/abyss.cc
	crawl-ref/source/acquire.cc
	crawl-ref/source/art-func.h
	crawl-ref/source/attitude-change.cc
	crawl-ref/source/beam.cc
	crawl-ref/source/behold.cc
	crawl-ref/source/command.cc
	crawl-ref/source/dat/database/randbook.txt
	crawl-ref/source/dat/descript/ability.txt
	crawl-ref/source/dat/descript/backgrounds.txt
	crawl-ref/source/dat/descript/branches.txt
	crawl-ref/source/dat/descript/cards.txt
	crawl-ref/source/dat/descript/commands.txt
	crawl-ref/source/dat/descript/features.txt
	crawl-ref/source/dat/descript/gods.txt
	crawl-ref/source/dat/descript/items.txt
	crawl-ref/source/dat/descript/monsters.txt
	crawl-ref/source/dat/descript/skills.txt
	crawl-ref/source/dat/descript/species.txt
	crawl-ref/source/dat/descript/spells.txt
	crawl-ref/source/dat/descript/unident.txt
	crawl-ref/source/dat/lua/autofight.lua
	crawl-ref/source/dat/tiles/Vera.ttf
	crawl-ref/source/dat/tiles/VeraMono.ttf
	crawl-ref/source/decks.cc
	crawl-ref/source/describe.cc
	crawl-ref/source/directn.cc
	crawl-ref/source/effects.cc
	crawl-ref/source/evoke.cc
	crawl-ref/source/fight.cc
	crawl-ref/source/fontwrapper-ft.cc
	crawl-ref/source/food.cc
	crawl-ref/source/godprayer.cc
	crawl-ref/source/invent.cc
	crawl-ref/source/item_use.cc
	crawl-ref/source/itemname.cc
	crawl-ref/source/itemprop.cc
	crawl-ref/source/items.cc
	crawl-ref/source/l_item.cc
	crawl-ref/source/libutil.cc
	crawl-ref/source/main.cc
	crawl-ref/source/melee_attack.cc
	crawl-ref/source/misc.cc
	crawl-ref/source/mon-abil.cc
	crawl-ref/source/mon-act.cc
	crawl-ref/source/mon-cast.cc
	crawl-ref/source/mon-data.h
	crawl-ref/source/mon-death.cc
	crawl-ref/source/mon-ench.cc
	crawl-ref/source/mon-project.cc
	crawl-ref/source/mon-stuff.cc
	crawl-ref/source/monster.cc
	crawl-ref/source/mutation-data.h
	crawl-ref/source/mutation.cc
	crawl-ref/source/newgame.cc
	crawl-ref/source/ouch.cc
	crawl-ref/source/output.cc
	crawl-ref/source/player-act.cc
	crawl-ref/source/player-equip.cc
	crawl-ref/source/player-stats.cc
	crawl-ref/source/player.cc
	crawl-ref/source/religion.cc
	crawl-ref/source/shopping.cc
	crawl-ref/source/skills.cc
	crawl-ref/source/spl-book.cc
	crawl-ref/source/spl-cast.cc
	crawl-ref/source/spl-clouds.cc
	crawl-ref/source/spl-damage.cc
	crawl-ref/source/spl-data.h
	crawl-ref/source/spl-goditem.cc
	crawl-ref/source/spl-miscast.cc
	crawl-ref/source/spl-selfench.cc
	crawl-ref/source/spl-util.cc
	crawl-ref/source/stairs.cc
	crawl-ref/source/startup.cc
	crawl-ref/source/status.cc
	crawl-ref/source/target.cc
	crawl-ref/source/teleport.cc
	crawl-ref/source/terrain.cc
	crawl-ref/source/tilereg-dgn.cc
	crawl-ref/source/tilereg-msg.cc
	crawl-ref/source/tilereg-stat.cc
	crawl-ref/source/transform.cc
	crawl-ref/source/traps.cc
	crawl-ref/source/view.cc
	crawl-ref/source/wiz-mon.cc
	crawl-ref/source/xom.cc

+++ crawl-ref/source/spl-clouds.cc
@@ -15,9 +15,7 @@
 #include "coord.h"
 #include "coordit.h"
 #include "env.h"
-#include "exercise.h"
 #include "fprop.h"
-#include "godconduct.h"
 #include "itemprop.h"
 #include "items.h"
 #include "libutil.h"
@@ -51,7 +49,7 @@ spret_type conjure_flame(int pow, const coord_def& where, bool fail)
 {
     // FIXME: This would be better handled by a flag to enforce max range.
     if (distance(where, you.pos()) > dist_range(spell_range(SPELL_CONJURE_FLAME,
-                                                      pow, true))
+                                                      pow))
         || !in_bounds(where))
     {
         mpr(gettext("That's too far away."));
@@ -62,9 +60,6 @@ spret_type conjure_flame(int pow, const coord_def& where, bool fail)
     {
         switch (grd(where))
         {
-        case DNGN_WAX_WALL:
-            mpr(gettext("The flames aren't hot enough to melt wax walls!"));
-            break;
         case DNGN_METAL_WALL:
             mpr(gettext("You can't ignite solid metal!"));
             break;
@@ -72,7 +67,7 @@ spret_type conjure_flame(int pow, const coord_def& where, bool fail)
             mpr(gettext("You can't ignite solid crystal!"));
             break;
         case DNGN_TREE:
-        case DNGN_SWAMP_TREE:
+        case DNGN_MANGROVE:
             fail_check();
             _burn_tree(where);
             return SPRET_SUCCESS;
@@ -138,7 +133,7 @@ spret_type stinking_cloud(int pow, bolt &beem, bool fail)
     beem.damage      = dice_def(1, 0);
     beem.hit         = 20;
     beem.glyph       = dchar_glyph(DCHAR_FIRED_ZAP);
-    beem.flavour     = BEAM_POTION_STINKING_CLOUD;
+    beem.flavour     = BEAM_POTION_MEPHITIC;
     beem.ench_power  = pow;
     beem.beam_source = MHITYOU;
     beem.thrower     = KILL_YOU;
@@ -242,9 +237,11 @@ void manage_fire_shield(int delay)
         you.duration[DUR_FIRE_SHIELD] = 0;
 
     // Remove fire clouds on top of you
-    if (env.cgrid(you.pos()) != EMPTY_CLOUD)
-        if (env.cloud[env.cgrid(you.pos())].type == CLOUD_FIRE)
-            delete_cloud_at(you.pos());
+    if (env.cgrid(you.pos()) != EMPTY_CLOUD
+        && env.cloud[env.cgrid(you.pos())].type == CLOUD_FIRE)
+    {
+        delete_cloud_at(you.pos());
+    }
 
     if (!you.duration[DUR_FIRE_SHIELD])
     {
@@ -328,7 +325,7 @@ static std::vector<int> _get_evaporate_result(int potion)
     case POT_PARALYSIS:
     case POT_CONFUSION:
     case POT_SLOWING:
-        beams.push_back(BEAM_POTION_STINKING_CLOUD);
+        beams.push_back(BEAM_POTION_MEPHITIC);
         break;
 
     case POT_WATER:
@@ -338,7 +335,7 @@ static std::vector<int> _get_evaporate_result(int potion)
 
     case POT_BLOOD:
     case POT_BLOOD_COAGULATED:
-        beams.push_back(BEAM_POTION_STINKING_CLOUD);
+        beams.push_back(BEAM_POTION_MEPHITIC);
         // deliberate fall through
     case POT_BERSERK_RAGE:
         beams.push_back(BEAM_POTION_FIRE);
@@ -362,7 +359,7 @@ static std::vector<int> _get_evaporate_result(int potion)
 
     default:
         beams.push_back(BEAM_POTION_FIRE);
-        beams.push_back(BEAM_POTION_STINKING_CLOUD);
+        beams.push_back(BEAM_POTION_MEPHITIC);
         beams.push_back(BEAM_POTION_COLD);
         beams.push_back(BEAM_POTION_POISON);
         beams.push_back(BEAM_POTION_BLUE_SMOKE);
@@ -379,14 +376,14 @@ static std::vector<int> _get_evaporate_result(int potion)
     {
         // handled in beam.cc
         clouds.push_back(CLOUD_FIRE);
-        clouds.push_back(CLOUD_STINK);
+        clouds.push_back(CLOUD_MEPHITIC);
         clouds.push_back(CLOUD_COLD);
         clouds.push_back(CLOUD_POISON);
         clouds.push_back(CLOUD_BLUE_SMOKE);
         clouds.push_back(CLOUD_STEAM);
     }
 
-    return (clouds);
+    return clouds;
 }
 
 // Returns a comma-separated list of all cloud types potentially created
@@ -438,7 +435,7 @@ spret_type cast_evaporate(int pow, bolt& beem, int pot_idx, bool fail)
     beem.ench_power = pow;               // used for duration only?
     beem.is_explosion = true;
 
-    beem.flavour    = BEAM_POTION_STINKING_CLOUD;
+    beem.flavour    = BEAM_POTION_MEPHITIC;
     beam_type tracer_flavour = BEAM_MMISSILE;
 
     switch (potion.sub_type)
@@ -469,7 +466,7 @@ spret_type cast_evaporate(int pow, bolt& beem, int pot_idx, bool fail)
         // fall through
     case POT_CONFUSION:
     case POT_SLOWING:
-        tracer_flavour = beem.flavour = BEAM_POTION_STINKING_CLOUD;
+        tracer_flavour = beem.flavour = BEAM_POTION_MEPHITIC;
         break;
 
     case POT_WATER:
@@ -523,7 +520,7 @@ spret_type cast_evaporate(int pow, bolt& beem, int pot_idx, bool fail)
         switch (random2(12))
         {
         case 0:   beem.flavour = BEAM_POTION_FIRE;            break;
-        case 1:   beem.flavour = BEAM_POTION_STINKING_CLOUD;  break;
+        case 1:   beem.flavour = BEAM_POTION_MEPHITIC;        break;
         case 2:   beem.flavour = BEAM_POTION_COLD;            break;
         case 3:   beem.flavour = BEAM_POTION_POISON;          break;
         case 4:   beem.flavour = BEAM_POTION_RANDOM;          break;
